solve/requirements.txt
pwntools
requests
playwright
